clearify view class
